Refactor cache index DAL SQL parser class for MySQL.
