+++ shardingsphere-test/shardingsphere-parser-test/src/main/java/org/apache/shardingsphere/test/sql/parser/parameterized/jaxb/cases/domain/statement/dal/CacheIndexStatementTestCase.java
@@ -18,8 +18,8 @@
 package org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.statement.dal;
 
 import lombok.Getter;
+import org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.segment.impl.index.ExpectedCacheTableIndex;
 import org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.segment.impl.index.ExpectedPartitionDefinition;
-import org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.segment.impl.index.ExpectedTableIndex;
 import org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.statement.SQLParserTestCase;
 
 import javax.xml.bind.annotation.XmlAttribute;
@@ -34,10 +34,10 @@ import java.util.List;
 public final class CacheIndexStatementTestCase extends SQLParserTestCase {
     
     @XmlElement(name = "table-index")
-    private final List<ExpectedTableIndex> tableIndexes = new LinkedList<>();
+    private final List<ExpectedCacheTableIndex> tableIndexes = new LinkedList<>();
     
     @XmlElement(name = "partition-definition")
-    private ExpectedPartitionDefinition partitions;
+    private ExpectedPartitionDefinition partitionDefinition;
     
     @XmlAttribute
     private String name;
